requirements.txt
ai2thor==2.5.0
autopep8==1.5.4
colour==0.1.5
commonmark==0.9.1
flake8==3.9.1
flake8-implicit-str-concat==0.2.0
flake8-use-fstring==1.3
importlib-metadata==4.13.0; python_version<"3.8"
isort==5.8.0
Jinja2<3.1
matplotlib==3.3.3
msgpack==1.0.0
numpyencoder==0.3.0
opencv-python==4.4.0.46; python_version<="3.9"
opencv-python==4.5.4.60; python_version>="3.10"
pep8-naming==0.12.1
pre-commit==2.12.1
pydantic==1.9.0
recommonmark==0.6.0
requests==2.24.0
scikit-image==0.18.3
Shapely==1.7.1; python_version<"3.10"
Shapely==1.8.0; python_version>="3.10"
Sphinx==3.2.1
sphinxcontrib-websupport==1.2.4
sphinx-markdown-builder==0.5.5
typeguard==2.13.2
git+https://github.com/sphinx-contrib/video.git
